+++ src/redux/modules/entities/dispatchMappings/selectors.js
@@ -23,7 +23,18 @@ export const getDispatchMappingValue = state =>
 
 export const selectDispatchMappingsFormInitialValues = state => {
   if (getSelectedEntity(state) === undefined) {
-    return new Map({ active: true, interactionField: 'customer', channelType: 'any' });
+    return new Map({ active: true, interactionField: 'customer', channelType: 'any', version: 'null' });
   }
-  return selectFormInitialValues(state);
+  return !selectFormInitialValues(state).get('version')
+    ? selectFormInitialValues(state).set('version', 'null')
+    : selectFormInitialValues(state);
 };
+
+export const getMappingValueMessage = state =>
+  `Enter${
+    ['voice', 'sms', 'messaging'].includes(currentChannelType(state)) && currentMappingValue(state) === 'contact-point'
+      ? ' an e.164 formatted number or SIP address'
+      : currentChannelType(state) === 'email' && currentMappingValue(state) === 'contact-point'
+        ? ' an email adress'
+        : 'a mapping value'
+  }`;