+++ src/redux/modules/entities/epics.js
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2015-2017 Serenova, LLC. All rights reserved.
+ * Copyright © 2015-2019 Serenova, LLC. All rights reserved.
  */
 
 import 'rxjs/add/operator/switchMap';
@@ -40,8 +40,6 @@ import {
 } from './selectors';
 
 import { getHasProficiencyFormValue } from './skills/selectors';
-
-import { selectNonReusableFlows, getFlowIdFormValue, selectVersionsFromFlow } from './flows/selectors';
 
 import { entitiesMetaData } from './metaData';
 
@@ -125,27 +123,6 @@ export const FocusOutboundIdentifiersValueFormField = (action$, store) =>
         getSelectedEntity(store.getState()) !== undefined
     )
     .map(a => touch(a.meta.form, 'value'));
-
-export const ClearMappingValue = (action$, store) =>
-  action$
-    .ofType('@@redux-form/CHANGE')
-    .filter(
-      a =>
-        a.meta.form.includes('dispatchMappings') &&
-        a.meta.field.includes('interactionField') &&
-        getSelectedEntity(store.getState()) !== undefined
-    )
-    .map(a => touch(a.meta.form, 'value'));
-
-export const TriggerFlowFetch = (action$, store) =>
-  action$
-    .ofType('@@redux-form/CHANGE')
-    .filter(a => a.meta.form.includes('dispatchMappings') && a.meta.field.includes('flowId'))
-    .map(a => ({
-      type: 'FETCH_DATA_FLOW',
-      entityName: 'flows',
-      id: a.payload
-    }));
 
 export const ToggleHasProficiencyFormField = (action$, store) =>
   action$
@@ -471,6 +448,7 @@ export const AddingListItems = (action$, store) =>
 export const addAndRemoveListItemEntities = (action$, store) =>
   action$
     .ofType('TOGGLE_LIST_ITEM_ENTITY')
+    .debounceTime(300)
     .map(a => ({
       ...a,
       entityName: getCurrentEntity(store.getState()),
@@ -830,67 +808,3 @@ export const DownloadCsv = (action$, store) =>
         })
         .catch(error => handleError(error, a))
     );
-
-export const changeDispatchMappingFlowId = (action$, store) =>
-  action$
-    .ofType('@@redux-form/REGISTER_FIELD', 'FETCH_DATA_FULFILLED')
-    .filter(
-      a =>
-        (a.meta !== undefined &&
-          a.meta.form === getSelectedEntityFormId(store.getState()) &&
-          a.payload.name === 'flowId' &&
-          selectNonReusableFlows(store.getState()).size &&
-          getSelectedEntityId(store.getState()) === 'create') ||
-        (getCurrentEntity(store.getState()) === 'dispatchMappings' &&
-          a.entityName === 'flows' &&
-          getSelectedEntityId(store.getState()) === 'create')
-    )
-    .map(a =>
-      //This change action is required as sometimes the flowId field has not been loaded and the field is left not initialized.
-      // That would not allow the user to submit, so the 'change' function will leave the field as 'touched' to Redux Form
-      change(getSelectedEntityFormId(store.getState()), 'flowId', selectNonReusableFlows(store.getState()).get(0).value)
-    );
-
-export const changeDispatchMappingVersion = (action$, store) =>
-  action$
-    .ofType('@@redux-form/REGISTER_FIELD', 'FETCH_DATA_FLOW_FULFILLED')
-    .filter(
-      a =>
-        (a.meta !== undefined &&
-          a.meta.form === getSelectedEntityFormId(store.getState()) &&
-          a.payload.name === 'version' &&
-          selectVersionsFromFlow(store.getState()) &&
-          selectVersionsFromFlow(store.getState()).length > 1 &&
-          getSelectedEntityId(store.getState()) === 'create') ||
-        (getCurrentEntity(store.getState()) === 'dispatchMappings' &&
-          a.entityName === 'flows' &&
-          getSelectedEntityId(store.getState()) === 'create')
-    )
-    .map(a =>
-      //This change action is required as sometimes the flowId field has not been loaded and the field is left not initialized.
-      // That would not allow the user to submit, so the 'change' function will leave the field as 'touched' to Redux Form
-      change(getSelectedEntityFormId(store.getState()), 'version', selectVersionsFromFlow(store.getState())[0].value)
-    );
-
-export const activateVersionsFromFlow = (action$, store) =>
-  action$
-    .ofType('@@redux-form/REGISTER_FIELD', 'FETCH_DATA_FULFILLED')
-    //We want to fetch the versions for the currently selected flow, but only if it has not been done before
-    .filter(
-      a =>
-        (a.meta !== undefined &&
-          a.meta.form === getSelectedEntityFormId(store.getState()) &&
-          a.payload.name === 'flowId' &&
-          selectNonReusableFlows(store.getState()).size &&
-          getSelectedEntityId(store.getState()) !== 'create' &&
-          !selectVersionsFromFlow(store.getState())) ||
-        (getCurrentEntity(store.getState()) === 'dispatchMappings' &&
-          a.entityName === 'flows' &&
-          getSelectedEntityId(store.getState()) !== 'create' &&
-          !selectVersionsFromFlow(store.getState()))
-    )
-    .map(a => ({
-      type: 'FETCH_DATA_FLOW',
-      entityName: 'flows',
-      id: getFlowIdFormValue(store.getState())
-    }));